pyproject.toml
[project]
name = "mjoshi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "baml-py>=0.71.1",
    "opencage>=3.0.4",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
